@@ -9,19 +9,18 @@
     "eslint": "./node_modules/.bin/eslint --quiet **/*.js"
   },
   "dependencies": {
-    "bluebird": "3.3.5",
-    "body-parser": "1.13.2",
+    "body-parser": "1.19.0",
     "config": "1.20.1",
     "cors": "2.7.1",
-    "express": "4.13.1",
+    "express": "4.17.1",
     "log4js": "0.6.35",
-    "soap": "0.14.0"
+    "mobiil-id-rest": "1.2.0",
+    "soap": "0.36.0"
   },
   "devDependencies": {
-    "chai": "3.5.0",
-    "eslint": "4.15.0",
-    "mocha": "2.4.5",
-    "supertest": "1.2.0",
-    "supertest-as-promised": "3.1.0"
+    "chai": "4.3.0",
+    "eslint": "7.20.0",
+    "mocha": "8.3.0",
+    "supertest": "6.1.3"
   }
 }
