@@ -1,42 +1,38 @@
 'use strict';
 
-var assert = require('chai').assert;
-var Response = require('../../../libs/models/Response');
+const assert = require('chai').assert;
+const Response = require('../../../libs/models/Response');
 
 suite('Response', function () {
 
     suite('Success', function () {
-        test('Construct - httpCode', function (done) {
-            var expected = {
+        test('Construct - httpCode', async function () {
+            const expected = {
                 status: {
                     code: 20000
                 }
             };
 
-            var response = new Response.Success(200);
+            const response = new Response.Success(200);
 
             assert.deepEqual(JSON.parse(JSON.stringify(response)), expected);
-
-            done();
         });
 
-        test('Construct - httpCode, statusMessage', function (done) {
-            var expected = {
+        test('Construct - httpCode, statusMessage', async function () {
+            const expected = {
                 status: {
                     code: 20000,
                     message: 'OK'
                 }
             };
 
-            var response = new Response.Success(200, 'OK');
+            const response = new Response.Success(200, 'OK');
 
             assert.deepEqual(JSON.parse(JSON.stringify(response)), expected);
-
-            done();
         });
 
-        test('Construct - httpCode, data', function (done) {
-            var expected = {
+        test('Construct - httpCode, data', async function () {
+            const expected = {
                 status: {
                     code: 20000
                 },
@@ -45,15 +41,13 @@ suite('Response', function () {
                 }
             };
 
-            var response = new Response.Success(200, {foo: 'bar'});
+            const response = new Response.Success(200, {foo: 'bar'});
 
             assert.deepEqual(JSON.parse(JSON.stringify(response)), expected);
-
-            done();
         });
 
-        test('Construct - httpCode, data (array)', function (done) {
-            var expected = {
+        test('Construct - httpCode, data (array)', async function () {
+            const expected = {
                 status: {
                     code: 20000
                 },
@@ -64,20 +58,18 @@ suite('Response', function () {
                 ]
             };
 
-            var response = new Response.Success(200, [
+            const response = new Response.Success(200, [
                 {
                     foo: 'bar'
                 }
             ]);
 
             assert.deepEqual(JSON.parse(JSON.stringify(response)), expected);
-
-            done();
         });
 
 
-        test('Construct - httpCode, statusMessage, data', function (done) {
-            var expected = {
+        test('Construct - httpCode, statusMessage, data', async function () {
+            const expected = {
                 status: {
                     code: 20000,
                     message: 'OK'
@@ -87,15 +79,13 @@ suite('Response', function () {
                 }
             };
 
-            var response = new Response.Success(200, 'OK', {foo: 'bar'});
+            const response = new Response.Success(200, 'OK', {foo: 'bar'});
 
             assert.deepEqual(JSON.parse(JSON.stringify(response)), expected);
-
-            done();
         });
 
-        test('Construct - httpCode, statusMessage, statusCode, data', function (done) {
-            var expected = {
+        test('Construct - httpCode, statusMessage, statusCode, data', async function () {
+            const expected = {
                 status: {
                     code: 20010,
                     message: 'OK'
@@ -105,46 +95,40 @@ suite('Response', function () {
                 }
             };
 
-            var response = new Response.Success(200, 'OK', 10, {foo: 'bar'});
+            const response = new Response.Success(200, 'OK', 10, {foo: 'bar'});
 
             assert.deepEqual(JSON.parse(JSON.stringify(response)), expected);
-
-            done();
         });
     });
 
     suite('Error', function () {
-        test('Construct - httpCode', function (done) {
-            var expected = {
+        test('Construct - httpCode', async function () {
+            const expected = {
                 status: {
                     code: 40000
                 }
             };
 
-            var response = new Response.Error(400);
+            const response = new Response.Error(400);
 
             assert.deepEqual(JSON.parse(JSON.stringify(response)), expected);
-
-            done();
         });
 
-        test('Construct - httpCode, statusMessage', function (done) {
-            var expected = {
+        test('Construct - httpCode, statusMessage', async function () {
+            const expected = {
                 status: {
                     code: 40000,
                     message: 'Bad Request'
                 }
             };
 
-            var response = new Response.Error(400, 'Bad Request');
+            const response = new Response.Error(400, 'Bad Request');
 
             assert.deepEqual(JSON.parse(JSON.stringify(response)), expected);
-
-            done();
         });
 
-        test('Construct - httpCode, errors', function (done) {
-            var expected = {
+        test('Construct - httpCode, errors', async function () {
+            const expected = {
                 status: {
                     code: 40000
                 },
@@ -153,15 +137,13 @@ suite('Response', function () {
                 }
             };
 
-            var response = new Response.Error(400, {field1: 'Error msg'});
+            const response = new Response.Error(400, {field1: 'Error msg'});
 
             assert.deepEqual(JSON.parse(JSON.stringify(response)), expected);
-
-            done();
         });
 
-        test('Construct - httpCode, statusMessage, errors', function (done) {
-            var expected = {
+        test('Construct - httpCode, statusMessage, errors', async function () {
+            const expected = {
                 status: {
                     code: 40000,
                     message: 'Bad Request'
@@ -171,15 +153,13 @@ suite('Response', function () {
                 }
             };
 
-            var response = new Response.Error(400, 'Bad Request', {field1: 'Error msg'});
+            const response = new Response.Error(400, 'Bad Request', {field1: 'Error msg'});
 
             assert.deepEqual(JSON.parse(JSON.stringify(response)), expected);
-
-            done();
         });
 
-        test('Construct - httpCode, statusMessage, statusCode, errors', function (done) {
-            var expected = {
+        test('Construct - httpCode, statusMessage, statusCode, errors', async function () {
+            const expected = {
                 status: {
                     code: 40010,
                     message: 'Bad Request'
@@ -189,26 +169,22 @@ suite('Response', function () {
                 }
             };
 
-            var response = new Response.Error(400, 'Bad Request', 10, {field1: 'Error msg'});
+            const response = new Response.Error(400, 'Bad Request', 10, {field1: 'Error msg'});
 
             assert.deepEqual(JSON.parse(JSON.stringify(response)), expected);
-
-            done();
         });
 
-        test('Construct - httpCode, statusMessage - 50000', function (done) {
-            var expected = {
+        test('Construct - httpCode, statusMessage - 50000', async function () {
+            const expected = {
                 status: {
                     code: 50000,
                     message: 'Internal Server Error'
                 }
             };
 
-            var response = new Response.Error(500, 'Internal Server Error');
+            const response = new Response.Error(500, 'Internal Server Error');
 
             assert.deepEqual(JSON.parse(JSON.stringify(response)), expected);
-
-            done();
         });
     });
 
